requirements.txt
openai==1.2.0
streamlit==1.18.0
altair<5
Pillow
scikit-learn